+++ docs/CHANGELOG.md
@@ -1,6 +1,11 @@
+## [v0.11.20]
+- Improved `phenocams.rois_mask_and_sum` to return the sum of pixel values for each channel and not treated as a single channel input image.
+  
+
 ## [v0.11.19] 
 - Improved `phenocams.overlay_polygons` to show the roi names at the centroid of the ROI polygon
 - Added `phenocams.rois_mask_and_sum` 
+- Added `phenocams.convert_rois_sums_to_single_dict`
   
 ## [v0.11.18]
 - Chasing a Heisenbug